pyproject.toml
[project]
name = "virusscan-web"
version = "0.1.0"
description = "Web application for virus scanning"
authors = [
    {name = "Felipe Almeida", email = "felipe@security.dev"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.4.0",
    "clamav-client==0.6.3",
    "django==4.2.18",
    "django-cotton>=1.5.2",
    "django-environ>=0.12.0",
    "django-ninja>=1.3.0",
    "psycopg[binary]>=3.2.4",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.4",
]
